DB_Setup/requirements.txt
numpy==1.14.5
mysql-connector-python==8.0.15
jsonschema==2.6.0
matplotlib==2.2.2
